refactor(database): add types split and namespaced module for TS migration
- Add lib/types/database.ts with public/modular types (Database, Statics, Reference, Query, etc.)
- Add lib/types/namespaced.ts with deprecated FirebaseDatabaseTypes namespace
- Add lib/types/internal.ts with DatabaseInternal and WithModularDeprecationArg
- Move FirebaseDatabaseModule and createModuleNamespace into lib/namespaced.js
- Slim lib/index.js to re-export from namespaced and modular
- Replace lib/index.d.ts with barrel re-exporting types and declare module augmentation
- Include database type sources in root tsconfig
