+++ packages/database/lib/namespaced.js
@@ -1,0 +1,204 @@
+/*
+ * Copyright (c) 2016-present Invertase Limited & Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this library except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import {
+  isAndroid,
+  isBoolean,
+  isNumber,
+  isString,
+  MODULAR_DEPRECATION_ARG,
+} from '@react-native-firebase/app/dist/module/common';
+import {
+  createModuleNamespace,
+  FirebaseModule,
+  getFirebaseRoot,
+} from '@react-native-firebase/app/dist/module/internal';
+import { setReactNativeModule } from '@react-native-firebase/app/dist/module/internal/nativeModule';
+import { createDeprecationProxy } from '@react-native-firebase/app/dist/module/common';
+import DatabaseReference from './DatabaseReference';
+import DatabaseStatics from './DatabaseStatics';
+import DatabaseTransaction from './DatabaseTransaction';
+import version from './version';
+import fallBackModule from './web/RNFBDatabaseModule';
+
+const namespace = 'database';
+
+const nativeModuleName = [
+  'RNFBDatabaseModule',
+  'RNFBDatabaseReferenceModule',
+  'RNFBDatabaseQueryModule',
+  'RNFBDatabaseOnDisconnectModule',
+  'RNFBDatabaseTransactionModule',
+];
+
+class FirebaseDatabaseModule extends FirebaseModule {
+  constructor(app, config, databaseUrl) {
+    super(app, config, databaseUrl);
+    this._serverTimeOffset = 0;
+    this._customUrlOrRegion = databaseUrl || this.app.options.databaseURL;
+    this._transaction = new DatabaseTransaction(this);
+    setTimeout(() => {
+      this._syncServerTimeOffset();
+    }, 100);
+  }
+
+  _syncServerTimeOffset() {
+    this.ref('.info/serverTimeOffset').on(
+      'value',
+      snapshot => {
+        this._serverTimeOffset = snapshot.val();
+      },
+      MODULAR_DEPRECATION_ARG,
+    );
+  }
+
+  getServerTime() {
+    return new Date(Date.now() + this._serverTimeOffset);
+  }
+
+  ref(path = '/') {
+    if (!isString(path)) {
+      throw new Error("firebase.app().database().ref(*) 'path' must be a string value.");
+    }
+
+    if (/[#$\[\]'?]/g.test(path)) {
+      throw new Error(
+        `Paths must be non-empty strings and can't contain #, $, [, ], ' or ? | path: ${path}`,
+      );
+    }
+
+    return createDeprecationProxy(new DatabaseReference(this, path));
+  }
+
+  refFromURL(url) {
+    if (!isString(url) || !url.startsWith('https://')) {
+      throw new Error(
+        "firebase.app().database().refFromURL(*) 'url' must be a valid database URL.",
+      );
+    }
+
+    if (!url.includes(this._customUrlOrRegion)) {
+      throw new Error(
+        `firebase.app().database().refFromURL(*) 'url' must be the same domain as the current instance (${this._customUrlOrRegion}). To use a different database domain, create a new Firebase instance.`,
+      );
+    }
+
+    let path = url.replace(this._customUrlOrRegion, '');
+    if (path.includes('?')) {
+      path = path.slice(0, path.indexOf('?'));
+    }
+
+    return createDeprecationProxy(new DatabaseReference(this, path || '/'));
+  }
+
+  goOnline() {
+    return this.native.goOnline();
+  }
+
+  goOffline() {
+    return this.native.goOffline();
+  }
+
+  setPersistenceEnabled(enabled) {
+    if (!isBoolean(enabled)) {
+      throw new Error(
+        "firebase.app().database().setPersistenceEnabled(*) 'enabled' must be a boolean value.",
+      );
+    }
+
+    return this.native.setPersistenceEnabled(enabled);
+  }
+
+  setLoggingEnabled(enabled) {
+    if (!isBoolean(enabled)) {
+      throw new Error(
+        "firebase.app().database().setLoggingEnabled(*) 'enabled' must be a boolean value.",
+      );
+    }
+
+    return this.native.setLoggingEnabled(enabled);
+  }
+
+  setPersistenceCacheSizeBytes(bytes) {
+    if (!isNumber(bytes)) {
+      throw new Error(
+        "firebase.app().database().setPersistenceCacheSizeBytes(*) 'bytes' must be a number value.",
+      );
+    }
+
+    if (bytes < 1048576) {
+      throw new Error(
+        "firebase.app().database().setPersistenceCacheSizeBytes(*) 'bytes' must be greater than 1048576 bytes (1MB).",
+      );
+    }
+
+    if (bytes > 104857600) {
+      throw new Error(
+        "firebase.app().database().setPersistenceCacheSizeBytes(*) 'bytes' must be less than 104857600 bytes (100MB).",
+      );
+    }
+
+    return this.native.setPersistenceCacheSizeBytes(bytes);
+  }
+
+  useEmulator(host, port) {
+    if (!host || !isString(host) || !port || !isNumber(port)) {
+      throw new Error('firebase.database().useEmulator() takes a non-empty host and port');
+    }
+    let _host = host;
+    const androidBypassEmulatorUrlRemap =
+      typeof this.firebaseJson.android_bypass_emulator_url_remap === 'boolean' &&
+      this.firebaseJson.android_bypass_emulator_url_remap;
+    if (!androidBypassEmulatorUrlRemap && isAndroid && _host) {
+      if (_host.startsWith('localhost')) {
+        _host = _host.replace('localhost', '10.0.2.2');
+        // eslint-disable-next-line no-console
+        console.log(
+          'Mapping database host "localhost" to "10.0.2.2" for android emulators. Use real IP on real devices. You can bypass this behaviour with "android_bypass_emulator_url_remap" flag.',
+        );
+      }
+      if (_host.startsWith('127.0.0.1')) {
+        _host = _host.replace('127.0.0.1', '10.0.2.2');
+        // eslint-disable-next-line no-console
+        console.log(
+          'Mapping database host "127.0.0.1" to "10.0.2.2" for android emulators. Use real IP on real devices. You can bypass this behaviour with "android_bypass_emulator_url_remap" flag.',
+        );
+      }
+    }
+    this.native.useEmulator(_host, port);
+    return [_host, port]; // undocumented return, just used to unit test android host remapping
+  }
+}
+
+export const SDK_VERSION = version;
+
+export default createModuleNamespace({
+  statics: DatabaseStatics,
+  version,
+  namespace,
+  nativeModuleName,
+  nativeEvents: ['database_transaction_event', 'database_sync_event'],
+  hasMultiAppSupport: true,
+  hasCustomUrlOrRegionSupport: true,
+  ModuleClass: FirebaseDatabaseModule,
+});
+
+export const firebase = getFirebaseRoot();
+
+for (let i = 0; i < nativeModuleName.length; i++) {
+  setReactNativeModule(nativeModuleName[i], fallBackModule);
+}
