+++ tsconfig.json
@@ -12,6 +12,9 @@
     "packages/crashlytics/lib/index.d.ts",
     "packages/crashlytics/lib/modular/index.d.ts",
     "packages/database/lib/index.d.ts",
+    "packages/database/lib/types/database.ts",
+    "packages/database/lib/types/namespaced.ts",
+    "packages/database/lib/types/internal.ts",
     "packages/database/lib/modular/index.d.ts",
     "packages/database/lib/modular/query.d.ts",
     "packages/database/lib/modular/transaction.d.ts",
